+++ src/constants/string-constnats.ts
@@ -17,7 +17,7 @@ export default {
         getBranches: 'git for-each-ref --format="{\'description\':\'%(objectname:short)\',\'label\':\'%(refname:short)\',\'current\':\'%(HEAD)\'}," refs/heads refs/remotes',
         getCurrentBranch: "git rev-parse --abbrev-ref HEAD",
         merge: (options: Array < string > , branchName ? : string, commitMessage?: string) => {
-            let command = "git merge " + (branchName || "");
+            let command = `git merge ${branchName || ""}`;
             if (options) {
                 options.forEach(option => {
                     if(option !== "m"){
@@ -26,13 +26,13 @@ export default {
                 });
             }
             if(commitMessage){
-                command += ' -m "' + commitMessage + '"';
+                command += ` -m "${commitMessage}"`;
             }
             return command;
         },
         stash: (stashCommand: string, includeOption:boolean = false, stashName ? : string) => {
             let option = includeOption ? '--pretty=format:"{\'detail\':\'%gd \u2022 %h \u2022 %cr\',\'label\':\'%s\',\'index\':\'%gd\'},"' : "",
-                command = "git stash " + stashCommand + option + (stashName || "");
+                command = `git stash ${stashCommand} ${option} ${stashName || ""}`;
             return command;
         }
     },
@@ -43,7 +43,7 @@ export default {
         DEBUG: "Debug"
     },
     error:
-        (errorWhile: string) => "Error while " + errorWhile,
+        (command: string) => `Error4 while ${command}`,
     timeForamt: {
         fullDate: "MM.DD.YYYY HH:mm:ss",
         hours: "HH:mm:ss"
@@ -61,10 +61,10 @@ export default {
     },
     success: {
         general: (operation, functionality) => {
-            return operation + " was successfully " + functionality;
+            return `${operation} was successfully ${functionality}`;
         },
         merge: (choosenBranch, currentBranch) => {
-            return "Branch '" + choosenBranch + "' was successfully merged into branch '" + currentBranch + "'";
+            return `Branch '${choosenBranch}' was successfully merged into branch '${currentBranch}'`;
         }
     }
 }